+++ totalsegmentator/libs.py
@@ -5,7 +5,9 @@ import sys
 import time
 import shutil
 import zipfile
+import subprocess
 from pathlib import Path
+import platform
 
 import requests
 import numpy as np
@@ -30,6 +32,16 @@ def nostdout(verbose=False):
         sys.stdout = save_stdout
     else:
         yield
+
+
+def get_config_dir():
+    if "TOTALSEG_WEIGHTS_PATH" in os.environ:
+        config_dir = Path(os.environ["TOTALSEG_WEIGHTS_PATH"]) / "nnUNet"
+    else:
+        # in docker container finding home not properly working therefore map to /tmp
+        home_path = Path("/tmp") if str(Path.home()) == "/" else Path.home()
+        config_dir = home_path / ".totalsegmentator/nnunet/results/nnUNet"
+    return config_dir
 
 
 # def download_url(url, save_path, chunk_size=128):
@@ -74,12 +86,7 @@ def download_url_and_unpack(url, config_dir):
 
 def download_pretrained_weights(task_id):
 
-    if "TOTALSEG_WEIGHTS_PATH" in os.environ:
-        config_dir = Path(os.environ["TOTALSEG_WEIGHTS_PATH"]) / "nnUNet"
-    else:
-        # in docker container finding home not properly working therefore map to /tmp
-        home_path = Path("/tmp") if str(Path.home()) == "/" else Path.home()
-        config_dir = home_path / ".totalsegmentator/nnunet/results/nnUNet"
+    config_dir = get_config_dir()
     (config_dir / "3d_fullres").mkdir(exist_ok=True, parents=True)
     (config_dir / "2d").mkdir(exist_ok=True, parents=True)
 
@@ -302,3 +309,59 @@ def check_if_shape_and_affine_identical(img_1, img_2):
         print("Shape out:")
         print(img_2.shape)
         print("WARNING: Output shape not equal to input shape. This should not happen.")
+
+
+def download_dcm2niix():
+    import urllib.request
+    print("  Downloading dcm2niix...")
+
+    if platform.system() == "Windows":
+        url = "https://github.com/rordenlab/dcm2niix/releases/latest/download/dcm2niix_win.zip"
+    elif platform.system() == "Darwin":  # Mac
+        if platform.machine().startswith("arm") or platform.machine().startswith("aarch"):  # arm
+            url = "https://github.com/rordenlab/dcm2niix/releases/latest/download/dcm2niix_mac_arm.pkg"
+        else:  # intel
+            url = "https://github.com/rordenlab/dcm2niix/releases/latest/download/dcm2niix_mac.zip"
+    elif platform.system() == "Linux":
+        url = "https://github.com/rordenlab/dcm2niix/releases/latest/download/dcm2niix_lnx.zip"
+    else:
+        raise ValueError("Unknown operating system. Can not download the right version of dcm2niix.")
+
+    config_dir = get_config_dir()
+
+    urllib.request.urlretrieve(url, config_dir / "dcm2niix.zip")
+    with zipfile.ZipFile(config_dir / "dcm2niix.zip", 'r') as zip_ref:
+        zip_ref.extractall(config_dir)
+
+    # Give execution permission to the script
+    os.chmod(config_dir / "dcm2niix", 0o755)
+
+    # Clean up
+    os.remove(config_dir / "dcm2niix.zip")
+    os.remove(config_dir / "dcm2niibatch")
+
+
+def dcm_to_nifti(input_path, output_path, verbose=False):
+    """
+    input_path: a directory of dicom slices
+    output_path: a nifti file path
+    """
+    verbose_str = "" if verbose else "> /dev/null"
+
+    config_dir = get_config_dir()
+    dcm2niix = config_dir / "dcm2niix"
+
+    if not dcm2niix.exists():
+        download_dcm2niix()
+
+    subprocess.call(f"{dcm2niix} -o {output_path.parent} -z y -f {output_path.name[:-7]} {input_path} {verbose_str}", shell=True)
+
+    nii_files = list(output_path.parent.glob("*.nii.gz"))
+    if len(nii_files) > 1:
+        print("WARNING: Dicom to nifti resulted in several nifti files. Only using first one.")
+        print([f.name for f in nii_files])
+        for nii_file in nii_files[1:]:
+            os.remove(nii_file)
+        # todo: have to rename first file to not contain any counter which is automatically added by dcm2niix
+
+    os.remove(str(output_path)[:-7] + ".json")